Fix reset() + unwinding double yield() segfault
Not only was reset() just pointing to run(), there were problems in
UnwindStack() which would make it seg fault in some cases. We were
reusing `yielded` to store the exception thrown for zombified fibers,
but that gets disposed aggressively (and rightfully so!) which leads to
us returning a disposed pointer in the case where you re-call yield()
from an unwinding fiber. And then in some cases v8 decided to
garbage collect it already and that's when the fun stops.