Hook pthreads earlier
It seems that if you have v8 compiled as a shared library instead of a
static library sometimes it can get initialized before coroutines.so, even
with LD_PRELOAD. I'm not really sure why this is the case, but this
means we need to run the hooks much earlier. So now everything that goes
through pthread_getspecific is fiber-local, instead of the way it used
to be where only keys created after coroutine.cc is initialized were
fiber-specific.