Misc refactor
This started off as a diff to keep a Locker active for each coroutine
even while not in use by v8, which would save the once-per-thread cost.
Turns out that cost isn't very high, and clearing out the threads is
better for garbage collection anyway. But the code cleaning up here was
pretty good so I'm putting this in.